Updating TrackViewer for double precision

+++ Source/ActivityRunner/Viewer3D/Debugging/DebugViewerForm.cs
@@ -2254,12 +2254,12 @@ namespace Orts.ActivityRunner.Viewer3D.Debugging
 
         protected static PointF PointFromLocation(in WorldLocation location)
         {
-            return new PointF(location.TileX * WorldLocation.TileSize + location.Location.X, location.TileZ * WorldLocation.TileSize + location.Location.Z);
+            return new PointF((float)(location.TileX * WorldLocation.TileSize + location.Location.X), (float)(location.TileZ * WorldLocation.TileSize + location.Location.Z));
         }
 
         protected static Vector2 VectorFromLocation(in WorldLocation location)
         {
-            return new Vector2(location.TileX * WorldLocation.TileSize + location.Location.X, location.TileZ * WorldLocation.TileSize + location.Location.Z);
+            return new Vector2((float)(location.TileX * WorldLocation.TileSize + location.Location.X), (float)(location.TileZ * WorldLocation.TileSize + location.Location.Z));
         }
     }
     #endregion
